Move pallet-bridge-dispatch types to primitives
